Update RequestController class
Added approveRequest and rejectRequest methods to the RequestController class. Replaced the type "int" with "Integer" in the RequestService class. Finally added checks in the  approveRequest and rejectRequest methods of the RequestService class to check if there is no request with the given id.

+++ eventManagementSystem/src/main/java/com/team3/eventManagementSystem/eventManagementSystem/service/RequestService.java
@@ -16,7 +16,7 @@ public class RequestService {
 	private List<ApprovalRequest> requestList = new ArrayList<>();
 	@Autowired
 	EventService eventService;
-	
+
 	/**
 	 * Adds a request to the List with the requests.
 	 * 
@@ -24,23 +24,21 @@ public class RequestService {
 	 */
 	public List<ApprovalRequest> createRequest(ApprovalRequest request) {
 		int newId = 1;
-		if(requestList.size() > 0) {
-			newId = requestList.get(requestList.size() - 1).getId() + 1; 
+		if (requestList.size() > 0) {
+			newId = requestList.get(requestList.size() - 1).getId() + 1;
 		}
 		request.setId(newId);
 		requestList.add(request);
 		return requestList;
 	}
-	
+
 	// Takes the id of a request and deletes the request
-	public List<ApprovalRequest> deleteRequest(int requestId) {
-		ApprovalRequest requestToDelete= requestList.stream()
-				.filter(r -> r.getId().equals(requestId))
-				.findFirst().orElse(null);
-		requestList.remove(requestToDelete); 
+	public List<ApprovalRequest> deleteRequest(Integer requestId) {
+		ApprovalRequest requestToDelete = requestList.stream().filter(r -> r.getId().equals(requestId)).findFirst()
+				.orElse(null);
+		requestList.remove(requestToDelete);
 		return requestList;
 	}
-
 
 	/**
 	 * Returns the list of the Requests
@@ -82,49 +80,52 @@ public class RequestService {
 		// Return the list of filtered requests
 		return filteredRequests;
 	}
+
 	// Takes the id of an ApprovalRequest and returns the ApprovalRequest
-	public ApprovalRequest getRequestById(int id) {
-		ApprovalRequest request= requestList.stream()
-				.filter(r -> r.getId().equals(id))
-				.findFirst().orElse(null);
+	public ApprovalRequest getRequestById(Integer id) {
+		ApprovalRequest request = requestList.stream().filter(r -> r.getId().equals(id)).findFirst().orElse(null);
 		return request;
 	}
-	
-	// Takes the id of a request, accepts it and 
+
+	// Takes the id of a request, accepts it and
 	// creates an event if the Request type is Create
 	// deletes it if it is Delete
-	public boolean approveRequest(int requestId) {
-	ApprovalRequest request= this.getRequestById(requestId);
-	Date today = new Date();
-	request.setClosedAt(today);
-	if (this.approvalRequestExists(request)) { 
-		if (request.getType().equals("create")) {
-			eventService.addEvent(request.getEvent()); // adds the event at the EventList
-		} else {// delete event
-			eventService.deleteEvent(request.getEvent().getId());
+	public List<ApprovalRequest> approveRequest(Integer requestId) {
+		ApprovalRequest request = this.getRequestById(requestId);
+		if (request != null) {
+			Date today = new Date();
+			request.setClosedAt(today);
+			if (this.approvalRequestExists(request)) {
+				if (request.getType().equals("create")) {
+					eventService.addEvent(request.getEvent()); // adds the event at the EventList
+				} else {// delete event
+					eventService.deleteEvent(request.getEvent().getId());
+				}
+				request.setStatus("accepted");
+
+			} else {
+				System.out.println(" Request not found");
+			}
 		}
-		request.setStatus("accepted");
-		return true;
-	} else {
-		System.out.println(" Request not found");
-		return false;
+		return this.getAllRequests();
 	}
-}
-	public boolean rejectRequest(int requestId) {
-		ApprovalRequest request= this.getRequestById(requestId);
-		Date today = new Date();
-		request.setClosedAt(today);
-		if (this.approvalRequestExists(request)) { // the request is at the list
-			request.setStatus("rejected");
-			// RequestService.deleteRequest(request); //Deletes the request if it is
-			// rejected
-			// Show rejection message
-			System.out.println(" The request for the event " + request.getEvent().getTitle() + " was rejected");
-			return true;
-		} else {
-			System.out.println(" The request doesn't exist");
-			return false;
+
+	public List<ApprovalRequest> rejectRequest(Integer requestId) {
+		ApprovalRequest request = this.getRequestById(requestId);
+		if (request != null) {
+			Date today = new Date();
+			request.setClosedAt(today);
+			if (this.approvalRequestExists(request)) { // the request is at the list
+				request.setStatus("rejected");
+				// RequestService.deleteRequest(request); //Deletes the request if it is
+				// rejected
+				// Show rejection message
+				System.out.println(" The request for the event " + request.getEvent().getTitle() + " was rejected");
+			} else {
+				System.out.println(" The request doesn't exist");
+			}
 		}
+		return this.getAllRequests();
 	}
 
 }
